Split collection and library into separate modules.
