+++ backend/src/jv2backend/app.py
@@ -10,7 +10,7 @@ import jv2backend.journalRoutes
 import jv2backend.generateRoutes
 import jv2backend.nexusRoutes
 import jv2backend.serverRoutes
-import jv2backend.journals
+import jv2backend.journalLibrary
 import jv2backend.io.journalLocator
 import jv2backend.io.journalGenerator
 import xml.etree.ElementTree as ElementTree
@@ -29,7 +29,7 @@ def create_app(inside_gunicorn: bool = True) -> Flask:
     # Create our main objects
     journalLocator = jv2backend.io.journalLocator.JournalLocator()
     journalGenerator = jv2backend.io.journalGenerator.JournalGenerator()
-    journalLibrary = jv2backend.journals.JournalLibrary({})
+    journalLibrary = jv2backend.journalLibrary.JournalLibrary({})
 
     # Register Flask routes
     jv2backend.serverRoutes.add_routes(app)
