+++ backend/src/jv2backend/io/journalGenerator.py
@@ -9,11 +9,11 @@ import xml.etree.ElementTree as ElementTree
 import hashlib
 
 import datetime
-from io import BytesIO
 from flask import make_response
 from jv2backend.requestData import RequestData, SourceType
 from jv2backend.utils import jsonify, url_join
-from jv2backend.journals import JournalLibrary, JournalCollection
+from jv2backend.journalLibrary import JournalLibrary
+from jv2backend.journalCollection import JournalCollection
 from jv2backend.journal import Journal
