+++ backend/src/jv2backend/journalCollection.py
@@ -129,36 +129,3 @@ class JournalCollection:
             results.update(matches)
 
         return results
-
-
-@dataclass
-class JournalLibrary:
-    """Defines one or more data source rootURL/directory and their associated
-    journal collections.
-    """
-    collections: typing.Dict[str, JournalCollection]
-
-    def __setitem__(self, key, value):
-        self.collections[key] = value
-
-    def __getitem__(self, key):
-        if key in self.collections:
-            return self.collections[key]
-        else:
-            return None
-
-    def __contains__(self, key):
-        return key in self.collections
-
-    def list(self):
-        """List contents of library"""
-        for c in self.collections:
-            logging.debug(f"Collection '{c}' contains "
-                          f"{len(self.collections[c].journalFiles)} "
-                          f"journal files:")
-            for j in self.collections[c].journalFiles:
-                if j.has_run_data():
-                    logging.debug(f"     {j.get_file_url()} "
-                                  f"({j.get_run_count()} run data)")
-                else:
-                    logging.debug(f"     {j.get_file_url()} (not yet loaded)")
