+++ frontend/data/sources.xml
@@ -1,7 +1,7 @@
 <xml>
 	<sources>
-		<source name="ISIS Archive" journalType="Network" journalRootUrl="http://data.isis.rl.ac.uk/journals" journalIndexFilename="journal_main.xml" instrumentSubdirs="true" runDataRootUrl="/archive"/>
 		<source name="Local" journalType="Cached" runDataRootUrl="/home/tris/src/jv2/TestData/Directories" dataOrganisation="Directory"/>
-		<source name="Local RB" journalType="Cached" instrumentSubdirs="true" runDataRootUrl="/home/ty11203/TestData/RB_Directories" dataOrganisation="RBNumber"/>
+		<source name="ISIS Archive" journalType="Network" journalRootUrl="http://data.isis.rl.ac.uk/journals" journalIndexFilename="journal_main.xml" instrumentSubdirs="true" runDataRootUrl="/archive"/>
+ŋ		<source name="Local RB" journalType="Cached" instrumentSubdirs="true" runDataRootUrl="/home/ty11203/TestData/RB_Directories" dataOrganisation="RBNumber"/>
 	</sources>
 </xml>
